Added: error handling and base model inclusion for categorization
